# api/docs/swagger.yaml
basePath: /services
definitions:
...
  api_internal_routes_middlewares.Unauthorized:
    properties:
      error:
        type: string
    type: object
  api_pkg_utils.ErrorResponse:
    properties:
      error:
        type: string
    required:
    - error
    type: object
...
  internal_services_occtl.ShowStatusResponse:
    properties:
      status:
        type: string
    type: object
  internal_services_occtl.UnBanIPRequest:
    properties:
      ip:
        type: string
    type: object
  internal_services_panel.CreateSiteConfigRequest:
    properties:
      google_captcha_secret_key:
        type: string
      google_captcha_site_key:
        type: string
    type: object
  internal_services_panel.GetFullPanelConfigResponse:
    properties:
      google_captcha_secret_key:
        type: string
      google_captcha_site_key:
        type: string
    type: object
  internal_services_panel.GetPanelConfigResponse:
    properties:
      google_captcha_site_key:
        type: string
      init:
        type: boolean
    type: object
  internal_services_panel.UpdateSiteConfigRequest:
    properties:
      google_captcha_secret_key:
        type: string
      google_captcha_site_key:
        type: string
    type: object
...
  occtl.IPBan:
    properties:
      _Since:
        type: string
      IP:
        type: string
      Since:
        type: string
    type: object
  occtl.IPBanPoints:
    properties:
      _Since:
        type: string
      IP:
        type: string
      Score:
        type: integer
      Since:
        type: string
    type: object
  occtl.IRoute:
    properties:
      Device:
        type: string
      ID:
        type: string
      IP:
        type: string
      Username:
        type: string
      iRoutes:
        type: string
      vhost:
        type: string
    type: object
  occtl.OcUser:
    properties:
      _Connected at:
        type: string
      _RX:
        type: string
      _TX:
        type: string
      Average RX:
        type: string
      Average TX:
        type: string
      Connected at:
        type: string
      DNS:
        items:
          type: string
        type: array
      DPD:
        type: string
      Device:
        type: string
      Full session:
        type: string
      Groupname:
        type: string
      Hostname:
        type: string
      IPv4:
        type: string
      KeepAlive:
        type: string
      Local Device IP:
        type: string
      Location:
        type: string
      MTU:
        type: string
      NBNS:
        items:
          type: string
        type: array
      P-t-P IPv4:
        type: string
      RX:
        type: string
      Remote IP:
        type: string
      Session:
        type: string
      Split-DNS-Domains:
        items:
          type: string
        type: array
      State:
        type: string
      TLS ciphersuite:
        type: string
      TX:
        type: string
      User-Agent:
        type: string
      Username:
        type: string
      iRoutes:
        items:
          type: string
        type: array
      raw_connected_at:
        type: integer
      vhost:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample Ocserv User management Api server.
  title: Ocserv User management Example Api
  version: "1.0"
paths:
  /api/v1/occtl/disconnect/:username:
    post:
      consumes:
      - application/json
      description: Disconnect User From Server and Close Session
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Disconnect User From Server
      tags:
      - Occtl
  /api/v1/occtl/ip_bans:
    get:
      consumes:
      - application/json
      description: List Of IP Bans
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Show IP bans with Points
        in: query
        name: points
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/occtl.IPBan'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: List Of IP Bans
      tags:
      - Occtl
  /api/v1/occtl/ip_bans/point:
    get:
      consumes:
      - application/json
      description: List Of IP Bans include points
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Show IP bans with Points
        in: query
        name: points
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/occtl.IPBanPoints'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: List Of IP Bans include points
      tags:
      - Occtl
  /api/v1/occtl/iroutes:
    get:
      consumes:
      - application/json
      description: Show IP Routes
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/occtl.IRoute'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Show IP Routes
      tags:
      - Occtl
  /api/v1/occtl/online:
    get:
      consumes:
      - application/json
      description: Show Online User List in Server
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/occtl.OcUser'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Online User List
      tags:
      - Occtl
  /api/v1/occtl/reload:
    post:
      consumes:
      - application/json
      description: Reload Server Configuration
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Reload Server
      tags:
      - Occtl
  /api/v1/occtl/status:
    get:
      consumes:
      - application/json
      description: Show Status Of Server State
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_services_occtl.ShowStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Show Status
      tags:
      - Occtl
  /api/v1/occtl/unban:
    post:
      consumes:
      - application/json
      description: Remove IP From Baned List
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: IP Address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_services_occtl.UnBanIPRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Unban IP
      tags:
      - Occtl
  /api/v1/occtl/users/:username:
    get:
      consumes:
      - application/json
      description: Show User Status
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv User Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/occtl.OcUser'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Show User Status
      tags:
      - Occtl
...
  /api/v1/panel/config:
    get:
      consumes:
      - application/json
      description: Get Panel Config
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_services_panel.GetFullPanelConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Get Panel Config
      tags:
      - Panel
    post:
      consumes:
      - application/json
      description: Create Panel Config initializing step
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: check secret key from file 'init_secret'
        in: query
        name: init_secret
        required: true
        type: string
      - description: site config data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_services_panel.CreateSiteConfigRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Create Panel Config
      tags:
      - Panel
    put:
      consumes:
      - application/json
      description: Update Panel Config in app after login step
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: site config data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_services_panel.UpdateSiteConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_internal_routes_middlewares.Unauthorized'
      summary: Update Panel Config
      tags:
      - Panel
  /api/v1/panel/config/init:
    get:
      consumes:
      - application/json
      description: Get Panel Config to discover init data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_services_panel.GetPanelConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_pkg_utils.ErrorResponse'
      summary: Get Panel Config with Init
      tags:
      - Panel
...
swagger: "2.0"